Flash module add
