+++ app/Modules/CSRFModule.php
@@ -10,22 +10,10 @@ class CSRFModule extends AModule
 {
     const MODULE_NAME = 'csrf';
 
-    public function checkRequireModule(array $arr = [])
-    {}
-
-    public function installModule()
-    {}
-
-    public function uninstallModule()
-    {}
-
     public function initialization(App $app)
     {
         parent::initialization($app);
     }
-
-    public function registerRoute()
-    {}
 
     public function registerDi()
     {
